+++ modules/platform/forge_runners/ec2_runners.tf
@@ -43,7 +43,6 @@ module "ec2_runners" {
       webhook_secret = aws_ssm_parameter.github_app_webhook_secret.value
     }
     runner_group_name = var.deployment_config.github.runner_group_name
-    scale_errors      = var.ec2_deployment_specs.scale_errors
     runner_specs      = var.ec2_deployment_specs.runner_specs
   }
 }